Moved timeline series into a subfolder.
