+++ js/masters/modules/timeline.src.js
@@ -11,4 +11,4 @@
  * License: www.highcharts.com/license
  */
 'use strict';
-import '../../Series/TimelineSeries.js';
+import '../../Series/Timeline/TimelineSeries.js';
